requirements.txt
streamlit>=1.28.0
openai>=1.3.0
langchain>=0.1.0
langchain-core>=0.1.0
langchain-openai>=0.0.2
langgraph>=0.0.40
pymongo>=4.6.0
python-dotenv>=1.0.0
pydantic>=2.5.0
regex>=2023.10.3
typing-extensions>=4.8.0